trying to add a rate for user

+++ app/src/main/java/ch/epfl/polychef/recipe/Rating.java
@@ -8,8 +8,8 @@ import java.util.Map;
 import ch.epfl.polychef.utils.Preconditions;
 
 public final class Rating implements Serializable {
-    private int ratingSum;
-    private Map<String, Integer> allRatings;
+    private double ratingSum;
+    private Map<String, Double> allRatings;
 
     /**
      * Constructs a new empty rating.
@@ -25,11 +25,11 @@ public final class Rating implements Serializable {
      * @param rate the rate given by the user, between 0 and 5
      * @return the rate that was overridden or -1 if there was no previous rate
      */
-    public int addRate(String userID, int rate){
+    public double addRate(String userID, double rate){
         Preconditions.checkArgument(0 <= rate && rate <= 5, "A rate's value should be between 0 and 5");
 
         if(allRatings.containsKey(userID)) {
-            int oldRate = allRatings.get(userID);
+            double oldRate = allRatings.get(userID);
             allRatings.replace(userID, rate);
             ratingSum = ratingSum - oldRate + rate;
             return oldRate;
@@ -54,11 +54,11 @@ public final class Rating implements Serializable {
         return String.format(Locale.ENGLISH,"%.2f", ratingAverage()) + "/5 stars by " + allRatings.size() + " users.\n";
     }
 
-    public int getRatingSum(){
+    public double getRatingSum(){
         return ratingSum;
     }
 
-    public Map<String, Integer> getAllRatings(){
+    public Map<String, Double> getAllRatings(){
         return allRatings;
     }
 }
